requirements.txt
hydra-core==1.1.1
mlagents==0.27.0
mlagents-envs==0.27.0
omegaconf==2.1.1
rich==10.16.1
torch==1.8.0
tqdm==4.62.3
wandb==0.12.9